+++ composeApp/src/commonMain/kotlin/pe/breaker/dkaviplay/data/mapper/InventarioMapper.kt
@@ -6,18 +6,18 @@ import pe.breaker.dkaviplay.data.remote.supabase.InventarioDTO
 
 fun InventarioEntity.toTable(): InventarioTable {
     return InventarioTable(
-        id = this.id.toLong(),
-        usuarioUid = this.usuarioUid.toLong(),
-        itemId = this.articuloId.toLong(),
+        id = this.id,
+        usuarioUid = this.usuarioUid,
+        itemId = this.articuloId,
         cantidad = this.cantidad.toLong()
     )
 }
 
 fun InventarioDTO.toEntity(): InventarioEntity {
     return InventarioEntity(
-        id = this.id.toInt(),
-        usuarioUid = this.userMovilId.toInt(),
-        articuloId = this.articulo.toInt(),
+        id = this.id,
+        usuarioUid = this.userMovilId,
+        articuloId = this.articulo,
         cantidad = this.cantidad
     )
 }
